requirements.txt
fastapi
uvicorn[standard]
playwright
beautifulsoup4
pdfminer.six
python-docx
sentence-transformers
faiss-cpu
SQLAlchemy
pydantic
requests
APScheduler
aiohttp
python-multipart
email-validator
jinja2
python-jose[cryptography]
passlib[bcrypt]
PyYAML
openai
redis
aioredis
celery
psutil
